重构了数据交换格式，使用protobuf传输序列化数据

+++ source/Style.h
@@ -3,19 +3,27 @@
 
 using namespace std;
 
-enum class TankStyle:uint8_t{ DEFAULT, BIG_DEFAULT, SMALL_DEFAULT };
-enum class BulletStyle:uint8_t{ DEFAULT, BIG_DEFAULT, SMALL_DEFAULT };
+enum class TankStyle : int
+{
+	DEFAULT,
+	BIG_DEFAULT,
+	SMALL_DEFAULT
+};
+enum class BulletStyle : int
+{
+	DEFAULT,
+	BIG_DEFAULT,
+	SMALL_DEFAULT
+};
 
 struct Style_info
 {
 	int width;
 	int height;
-	Style_info(int width, int height) :
-		width(width), height(height){};
+	Style_info(int width, int height) : width(width), height(height){};
 };
 
-
-extern map<TankStyle, Style_info*> Tank_Style_info;
-extern map<BulletStyle, Style_info*> Bullet_Style_info;
+extern map<TankStyle, Style_info *> Tank_Style_info;
+extern map<BulletStyle, Style_info *> Bullet_Style_info;
 
 void Init_Style();
